chore(.NET): update hierarchy examples
